Fixed a bunch of UI bugs
Removed the -n argument. Now just generating a name based on
the search term, season, and episode

Automatically selects the first search result if it is the
only one.

Only prints season numbers if it needs user input

Only prints episode numbers if it needs user input

Prints what season and episode the program is selecting.
This is for more user feedback on what's happening.

Changed the temporary file name to the same as the generated
file name. Makes everything simpler and the code looks cleaner

Standardized the string asking what program to run.

Made vlc the default program to run if nothing entered.

Waits for the file to come into existance and have some data in
it before running the video program.